Enabled bower publish only for tags [ci skip]

+++ publish.sh
@@ -1,15 +1,18 @@
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "Travis CI"
+if [ ! -z "$TRAVIS_TAG" ]; then
+    git config --global user.email "travis@travis-ci.org"
+    git config --global user.name "Travis CI"
 
-git clone https://github.com/serban-petrescu/packaged-ui5-auth.git
+    git clone https://github.com/serban-petrescu/packaged-ui5-auth.git
 
-cp -r dist/* packaged-ui5-auth/
-cd packaged-ui5-auth
+    cp -r dist/* packaged-ui5-auth/
+    cd packaged-ui5-auth
 
-git remote rm origin
-git remote add origin https://serban-petrescu:${GITHUB_TOKEN}@github.com/serban-petrescu/packaged-ui5-auth.git > /dev/null 2>&1
+    git remote rm origin
+    git remote add origin https://serban-petrescu:${GITHUB_TOKEN}@github.com/serban-petrescu/packaged-ui5-auth.git > /dev/null 2>&1
 
-git add . *.*
-git commit --message "Travis build: $TRAVIS_BUILD_NUMBER"
+    git add . *.*
+    git commit --message "Travis build: $TRAVIS_BUILD_NUMBER"
+    git tag -a "$TRAVIS_TAG" -m "$TRAVIS_TAG"
 
-git push --follow-tags --quiet --set-upstream origin master
+    git push --follow-tags --quiet --set-upstream origin master
+fi
